add unit tests

+++ apps/api/src/database/database.ts
@@ -1,4 +1,4 @@
-import { Mongoose, Types, connect } from 'mongoose'
+import { Mongoose, Types, connect, disconnect } from 'mongoose'
 import { Question } from './schemas/question.schema'
 import { questionsData } from './data/questions'
 import { IQuestionDTO } from '@orthographiq/shared'
@@ -25,25 +25,34 @@ export class MongoDB {
 		}
 	}
 
+	static async disconnect(): Promise<void> {
+		if (MongoDB.db) {
+			await disconnect()
+			MongoDB.db = undefined
+		}
+	}
+
 	/**
 	 * Creates the question in the DB
 	 */
 	private static async loadData(): Promise<void> {
-		await Question.countDocuments()
-			.exec()
-			.then(async (count: number) => {
-				if (count !== questionsData.length) {
-					await Question.deleteMany()
-					await Question.create(
-						questionsData.map(
-							(question: IQuestionDTO) =>
-								new Question({
-									...question,
-									_id: new Types.ObjectId(),
-								})
+		if (config.env !== 'test') {
+			await Question.countDocuments()
+				.exec()
+				.then(async (count: number) => {
+					if (count !== questionsData.length) {
+						await Question.deleteMany()
+						await Question.create(
+							questionsData.map(
+								(question: IQuestionDTO) =>
+									new Question({
+										...question,
+										_id: new Types.ObjectId(),
+									})
+							)
 						)
-					)
-				}
-			})
+					}
+				})
+		}
 	}
 }
